+++ packages/terra-form-input/CHANGELOG.md
@@ -4,6 +4,11 @@ ChangeLog
 Unreleased
 ----------
 
+3.0.0 - (March 31, 2020)
+------------------
+### Breaking Changes
+* `label` prop's type changed from node to string.
+
 2.36.0 - (March 10, 2020)
 ------------------
 ### Changed